+++ packages/core/transport/src/index.ts
@@ -1,5 +1,5 @@
-export * from "./BaseTransport.js";
-export * from "./types.js";
 export { LogLevel } from "@loglayer/shared";
-export * from "./test-utils.js";
+export * from "./BaseTransport.js";
 export * from "./LoggerlessTransport.js";
+export * from "./test-utils.js";
+export * from "./types.js";