+++ packages/transports/opentelemetry/src/utils.ts
@@ -2,7 +2,7 @@
 // https://github.com/open-telemetry/opentelemetry-js-contrib/blob/main/packages/winston-transport/src/OpenTelemetryTransportV3.ts
 // Which is Copyright The OpenTelemetry Authors and licensed under the Apache License, Version 2.0.
 
-import { type LogAttributes, type LogRecord, type Logger, SeverityNumber } from "@opentelemetry/api-logs";
+import { type LogAttributes, type Logger, type LogRecord, SeverityNumber } from "@opentelemetry/api-logs";
 
 const logLayerLevels: Record<string, number> = {
   fatal: SeverityNumber.FATAL,
@@ -21,7 +21,7 @@ export function emitLogRecord(record: Record<string, any>, logger: Logger): void
   const { message, level, ...splat } = record;
   const attributes: LogAttributes = {};
   for (const key in splat) {
-    if (Object.prototype.hasOwnProperty.call(splat, key)) {
+    if (Object.hasOwn(splat, key)) {
       attributes[key] = splat[key];
     }
   }